+++ mhw_analysis/systems/src/build.h
@@ -4,8 +4,8 @@
 
 #include <stdbool.h>
 
-int column_to_row_major_index(int k, int nr, int nc);
-void flat_row_major_indices(int k, int nr, int nc, int *i, int *j);
+int first_pass(bool ***cube, int *shape, bool upper_left);
+void lunion(int *parent, int x, int y);
 
-#endif // _BSPLINE_H_
+#endif // _BUILD_H_
